# openapi.yaml
openapi: 3.0.3
info:
  title: API Ayuntamiento
  description: API REST para la gestión de noticias, eventos y reservas del local municipal.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Desarrollo local
...
paths:
...
  /noticias:
    get:
      tags: [Noticias]
      summary: Listar todas las noticias
      responses:
        '200':
          description: Lista de noticias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Noticia'

    post:
      tags: [Noticias]
      summary: Crear una noticia
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/NoticiaInput'
      responses:
        '201':
          description: Noticia creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Noticia'
        '400':
          $ref: '#/components/responses/BadRequest'
  /noticias/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'

    get:
      tags: [Noticias]
      summary: Obtener una noticia por ID
      responses:
        '200':
          description: Noticia encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Noticia'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Noticias]
      summary: Actualizar una noticia
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/NoticiaInput'
      responses:
        '200':
          description: Noticia actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Noticia'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Noticias]
      summary: Eliminar una noticia
      responses:
        '204':
          description: Noticia eliminada
        '404':
          $ref: '#/components/responses/NotFound'

  # ─────────────────────────────────────────
  # EVENTOS
  # ─────────────────────────────────────────
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        example: 1
  responses:
    BadRequest:
      description: Datos de entrada inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Noticia:
      type: object
      properties:
        id:
          type: integer
          example: 1
        titulo:
          type: string
          example: 'Corte de agua el lunes'
        descripcion:
          type: string
          example: 'El suministro estará cortado de 9h a 14h en la calle Mayor.'
        imagen:
          type: string
          nullable: true
          example: '/uploads/noticias/corte-agua.jpg'
        fechaSubida:
          type: string
          format: date-time
          example: '2026-06-25T10:00:00Z'
    NoticiaInput:
      type: object
      required: [titulo, descripcion]
      properties:
        titulo:
          type: string
          example: 'Corte de agua el lunes'
        descripcion:
          type: string
          example: 'El suministro estará cortado de 9h a 14h en la calle Mayor.'
        imagen:
          type: string
          format: binary
          description: Archivo de imagen (opcional)

    # ── Eventos ────────────────────────────
...
    Error:
      type: object
      properties:
        mensaje:
          type: string
          example: 'Recurso no encontrado'